# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Tayooli ERP API
  version: "5.0.0"
  description: |
    REST API for the Tayooli ERP Go backend.
    All endpoints under `/api/v1` require a signed HS256 JWT that carries a
    `tenant_id` claim (UUID).  The middleware extracts the claim and enforces
    dual-layer tenant isolation: an application-level WHERE clause plus a
    PostgreSQL `SET LOCAL app.current_tenant_id` that activates RLS policies.

    **Tenant ID handling**: `tenant_id` is always derived from the JWT claim.
    Any `tenant_id` field in a request body is silently ignored.  Never pass
    tenant identity through the request body.
  contact:
    name: Tayooli Engineering
    email: sppgcurugsukabakti@gmail.com
servers:
  - url: http://localhost:8081
    description: Local development
  - url: https://api.tayooli.id
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256-signed JWT.  Required claims:
        - `sub` — user UUID
        - `tenant_id` — tenant UUID (used for all tenant isolation logic)
        - `exp` — expiry timestamp (required; token is rejected if missing)
  schemas:
...
    Invoice:
      type: object
      required:
        - id
        - tenant_id
        - vendor_id
        - invoice_number
        - amount
        - currency
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        tenant_id:
          type: string
          format: uuid
          example: "b1a2c3d4-e5f6-7890-abcd-ef1234567890"
        vendor_id:
          type: string
          example: "VENDOR-001"
        invoice_number:
          type: string
          example: "INV-2026-00042"
        amount:
          type: string
          example: "1500000.0000"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: "IDR"
        status:
          type: string
          enum: [pending, ai_processing, ai_processed, ai_failed, approved, rejected]
          example: "pending"
        ai_confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          nullable: true
          description: OCR confidence score (0.0–1.0). Null until AI worker processes the document.
          example: 0.94
        extracted_text:
          type: string
          nullable: true
          description: Raw text extracted by Tesseract OCR.
          example: "PT Maju Bersama\nInvoice No: INV-2026-00042\nTotal: Rp 1.500.000"
        ocr_processed_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when OCR processing completed.
          example: "2026-06-27T09:15:42Z"
        due_date:
          type: string
          format: date-time
          nullable: true
          example: "2026-07-31T00:00:00Z"
        created_at:
          type: string
          format: date-time
          example: "2026-06-26T08:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-06-26T08:00:00Z"

    # ── Purchase Order ─────────────────────────────────────────────────────────
...
    CreateInvoiceRequest:
      type: object
      required:
        - vendor_id
        - invoice_number
        - amount
        - currency
      properties:
        vendor_id:
          type: string
          example: "VENDOR-001"
        invoice_number:
          type: string
          example: "INV-2026-00042"
        amount:
          type: string
          pattern: '^\d{1,16}(\.\d{1,4})?$'
          description: 'Decimal amount as string, e.g. "12345.6789". Max 4 decimal places. Max value 9999999999999999.9999'
          example: "1500000.0000"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: "IDR"
        due_date:
          type: string
          format: date-time
          nullable: true
          example: "2026-07-31T00:00:00Z"
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "invoice not found"

# ── Global security ────────────────────────────────────────────────────────────
security:
  - BearerAuth: []

# ── Paths ──────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/invoices:

    get:
      operationId: listInvoices
      summary: List invoices
      description: |
        Returns all invoices belonging to the authenticated tenant, ordered by
        `created_at DESC`.  The tenant is determined exclusively from the JWT
        `tenant_id` claim — no query parameter is accepted.
      tags: [invoices]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Paginated list of invoices (currently returns all records).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
              example:
                - id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  tenant_id: "b1a2c3d4-e5f6-7890-abcd-ef1234567890"
                  vendor_id: "VENDOR-001"
                  invoice_number: "INV-2026-00042"
                  amount: "1500000.0000"
                  currency: "IDR"
                  status: "ai_processed"
                  ai_confidence_score: 0.94
                  extracted_text: "PT Maju Bersama\nInvoice No: INV-2026-00042\nTotal: Rp 1.500.000"
                  ocr_processed_at: "2026-06-27T09:15:42Z"
                  due_date: "2026-07-31T00:00:00Z"
                  created_at: "2026-06-27T09:00:00Z"
                  updated_at: "2026-06-27T09:15:42Z"
        "401":
          description: Missing, expired, or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_header:
                  value: { "error": "missing authorization header" }
                invalid_token:
                  value: { "error": "invalid token" }
                missing_tenant:
                  value: { "error": "missing tenant_id in token" }
        "500":
          description: Unexpected server-side error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "internal server error"

    post:
      operationId: createInvoice
      summary: Create an invoice
      description: |
        Creates a new invoice in `pending` status for the authenticated tenant.
        On success a `invoice.created` event is published to Kafka (key = invoice UUID).
        Publishing is best-effort — a Kafka failure does NOT roll back the database write.

        **Validation rules** (enforced by the use-case layer):
        - `vendor_id` must not be empty.
        - `invoice_number` must not be empty.
        - `amount` must be greater than 0.
        - `currency` must not be empty (3-letter ISO 4217 recommended).
      tags: [invoices]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
            example:
              vendor_id: "VENDOR-001"
              invoice_number: "INV-2026-00042"
              amount: "1500000.0000"
              currency: "IDR"
              due_date: "2026-07-31T00:00:00Z"
      responses:
        "201":
          description: Invoice created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
              example:
                id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                tenant_id: "b1a2c3d4-e5f6-7890-abcd-ef1234567890"
                vendor_id: "VENDOR-001"
                invoice_number: "INV-2026-00042"
                amount: "1500000.0000"
                currency: "IDR"
                status: "pending"
                ai_confidence_score: null
                extracted_text: null
                ocr_processed_at: null
                due_date: "2026-07-31T00:00:00Z"
                created_at: "2026-06-27T09:00:00Z"
                updated_at: "2026-06-27T09:00:00Z"
        "400":
          description: Malformed JSON body (cannot be decoded).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "invalid request body"
        "401":
          description: Missing, expired, or invalid JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "unauthorized"
        "422":
          description: |
            Request body is valid JSON but fails domain validation
            (empty fields, non-positive amount, etc.).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "invalid invoice data"
        "500":
          description: Unexpected server-side error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "internal server error"
...